backend/requirements.txt
aiosmtpd==1.4.6
alembic==1.15.2
amqp==5.3.1
apispec==6.8.2
atpublic==6.0.1
attrs==25.3.0
billiard==4.2.1
black==25.1.0
blinker==1.9.0
cachelib==0.13.0
celery==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.1.0
Flask-Caching==2.3.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
flask-smorest==0.46.1
Flask-SQLAlchemy==3.1.1
flower==2.0.1
greenlet==3.2.2
gunicorn==23.0.0
humanize==4.12.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.3
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
prometheus_client==0.22.0
prompt_toolkit==3.0.51
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==5.3.0
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.40
tornado==6.5
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
webargs==8.7.0
Werkzeug==3.1.3
